Re-do from start.
Basically removed the "typed streams" from low level and left it to be arranged
using signal names. This reduces complexity at minor cost of "escaping" reserved
names so that they are usable. Using "false" and "true" as type names for "boolean"
together with names registry allows it to be almost as compact as typed indicator
streams.

This change will also simplify testing because "described" format will be just
one API laid over already tested API and removes the "optional" behavior from
primary implementation. This also introduced a hell lot of complexity.

Removed the concept of "indicator format" as an independent entity and moved it
conceptualy to implementation level. The indicator format failed to be much help
due to complexity of it's state management and incompatiblity especially with
JSON format.

The properly layered abstract methods API seems to be more profitable approach
allowing less messy implementation.